+++ R/epoxy_style_html.R
@@ -1,4 +1,3 @@
-
 #' Concise syntax for expressions inside HTML elements
 #'
 #' @description